--- hydra-topic-service/src/main/java/com/revature/HydraTopicServiceApplication.java
@@ -1,36 +1,0 @@
-package com.revature;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.client.circuitbreaker.EnableCircuitBreaker;
-import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
-import org.springframework.context.annotation.Bean;
-
-import springfox.documentation.builders.PathSelectors;
-import springfox.documentation.builders.RequestHandlerSelectors;
-import springfox.documentation.spi.DocumentationType;
-import springfox.documentation.spring.web.plugins.Docket;
-import springfox.documentation.swagger2.annotations.EnableSwagger2;
-
-@EnableSwagger2
-@EnableCircuitBreaker
-@EnableEurekaClient
-@SpringBootApplication
-public class HydraTopicServiceApplication {
-	public static void main(String[] args) {
-		SpringApplication.run(HydraTopicServiceApplication.class, args);
-	}
-
-	/**
-	 * Used to set up Swagger documentation
-	 * 
-	 * @author Jennifer Fox - Batch Matt 1802
-	 * 
-	 * @return Docket object for Swagger to use
-	 */
-	@Bean
-	public Docket api() {
-		return new Docket(DocumentationType.SWAGGER_2).select().apis(RequestHandlerSelectors.any())
-				.paths(PathSelectors.any()).build();
-	}
-}
