--- hydra-topic-service/src/main/resources/bootstrap.yml
@@ -1,24 +1,0 @@
-server:
-  port: 5300
-
-spring:
-  application:
-    name: hydra-topic-service
-  cloud:
-    config:
-      uri: http://localhost:8888
-  datasource:
-    url: jdbc:h2:mem:testTopic
-    username: roll
-    password: tide
-    driver-class-name: org.h2.Driver
-  jpa:
-    database-platform: org.hibernate.dialect.H2Dialect
-    show-sql: false
-    properties:
-      hibernate:
-        format-sql: false
-  h2:
-    console:
-      enabled: true
-      path: /h2
